app/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
python-dotenv
PyJWT
pydantic
yfinance